requirements.txt
pillow
python-dotenv
google-genai
requests
